Reorganize views.py: annotations, comments, clients into controllers
